scraper/requirements.txt
apify-client==1.7.1
requests==2.31.0
python-dotenv==1.0.0